Migrate components to TS.
