+++ graylog2-web-interface/src/components/event-definitions/event-definition-types/bindings.ts
